+++ src/tpl.js
@@ -22,8 +22,8 @@ export function corner (data) {
 /**
  * Render main content
  */
-export function main (tpl) {
-  const aside = `${tpl}<aside class="sidebar"></aside>`
+export function main () {
+  const aside = `${toggle()}<aside class="sidebar"></aside>`
 
   return (isMobile() ? `${aside}<main>` : `<main>${aside}`) +
     `<section class="content">
@@ -74,4 +74,8 @@ export function tree (toc, tpl = '') {
 
 export function helper (className, content) {
   return `<p class="${className}">${content.slice(5).trim()}</p>`
+}
+
+export function theme (color) {
+  return `<style>:root{--theme-color: ${color};}</style>`
 }